# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Description of your API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/routes": {
            "get": {
                "tags": [
                    "Routes"
                ],
                "summary": "Get a list of routes",
                "operationId": "25d3d0d0f5d12b021dc6272e30293e95",
                "responses": {
                    "200": {
                        "description": "List of routes.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Route"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Routes"
                ],
                "summary": "Create a new route",
                "operationId": "b33c4c1179bc9292122b3d06edfeaff1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "/schemas/RouteRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Route created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/schemas/Route"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                }
            }
        },
        "/api/routes/{id}": {
            "get": {
                "tags": [
                    "Routes"
                ],
                "summary": "Get a specific route",
                "operationId": "e2a8d29a88e0b4b65e46b742ac68f79e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Route retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/schemas/Route"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Route not found."
                    }
                }
            },
            "put": {
                "tags": [
                    "Routes"
                ],
                "summary": "Update a specific route",
                "operationId": "deeadf381bc7b3d6471d4c17c0cfe8ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "/schemas/RouteRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Route updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/schemas/Route"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You are not authorized to update this route."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Routes"
                ],
                "summary": "Delete a specific route",
                "operationId": "90c67fe2765db8d58d5efc5a24316169",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Route deleted successfully."
                    },
                    "403": {
                        "description": "You are not authorized to delete this route."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Route": {
                "required": [
                    "id",
                    "name",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Route 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description of the route"
                    },
                    "start_location": {
                        "type": "string",
                        "example": "Start Location"
                    },
                    "end_location": {
                        "type": "string",
                        "example": "End Location"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}